% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_burgio_published_data.R
\name{get_burgio_published_data}
\alias{get_burgio_published_data}
\title{Download originally published data from Burgio et al. (2018)}
\usage{
get_burgio_published_data(dir = "data-raw/temp")
}
\arguments{
\item{dir}{Where to save the .zip and all unpacked files. Defaults to /data-raw/temp/}
}
\value{
Downloads and unzips .zip archive from the Burgio published data at Figshare. Saves to dir
}
\description{
Helper function for downloading the Burgio data directly to file
}
