% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ptdb.R
\name{get_ptdb}
\alias{get_ptdb}
\title{Retrieve the Parrot Traits Database}
\usage{
get_ptdb()
}
\description{
Provides two options for retrieving and pulling into memory the Parrot Traits Database (PTDB). The first and default option is to use the data stored in the package (`r data(ptb)`).
}
