% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supertree}
\alias{supertree}
\title{Originally published Parrot phylogenetic tree as published at Burgio et al. (2018)}
\format{
A list containing (1) a Phylogenetic tree of class "phylo" with 413 tips and 312 internal nodes; tip labels are scientific names of Parrot species as published in Burgio et al. (2018); (2) node.dates, a data.frame of 156 rows and 3 columns.
}
\source{
\url{https://ndownloader.figshare.com/articles/5877324?private_link=6cdf8cf00793deab7ba6}
}
\usage{
data("supertree")
}
\description{
Description as published in [FigShare](https://ndownloader.figshare.com/articles/5877324?private_link=6cdf8cf00793deab7ba6):
"We present a complete dataset from the literature on functional traits including morphological measurements, dietary information, foraging strategy, and foraging location for all 398 extant species of parrots. The morphological measurements include: mass as well as total length, wing chord, culmen, tarsus, and tail length. The diet data describe whether each species is known to consume particular food items (e.g. nectar, berries, and carrion), foraging strategy data describes how each species captures or accesses food, and foraging location describes the habitat from which each species finds food (e.g. ground, canopy, and subcanopy). We also present a time-calibrated phylogenetic supertree that contains all 398 extant species as well as 15 extinct species (413 total species)."
}
\keyword{datasets}
