% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits}
\alias{traits}
\title{Parrot Trait Database}
\format{
A data frame with 414 rows (species) and 180 variables (traits and taxonomic information:
\describe{
}
}
\source{
Kevin R. Burgio personal communication
}
\usage{
data(traits)
}
\description{
A comprehensive and curated dataset containing all known and credible traits for parrot species worldwide.
Taxonomic information used to define and identify individual species is _____
}
\keyword{datasets}
